refactor(workflow): isolate runtime job state
